requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
PyPDF2
python-dotenv